+++ src/common/types/stat-types.ts
@@ -22,7 +22,7 @@ export type FeatherStatName = Percentage<ATK>;
 export type SandsStatName = Percentage<BasicStatName> | EnergyRecharge | ElementalMastery;
 export type GobletStatName = Percentage<BasicStatName> | ElementalDMGBonus | PhysicalDMGBonus | ElementalMastery;
 export type CircletStatName = Percentage<BasicStatName> | Crits | HealingBonus | ElementalMastery;
-export type WeaponStatName = BasicStatName | PhysicalDMGBonus | Crits | ElementalMastery | EnergyRecharge
+export type WeaponStatName = BasicStatName | PhysicalDMGBonus  | ElementalMastery | EnergyRecharge | Crits
 
 export type StatName = 
   | Percentage<BasicStatName>