+++ src/components/common/media/icons/Rarity/Rarity.tsx
@@ -4,12 +4,13 @@ import { classNames, rarityString } from "@/common/functions/strings";
 
 type Props = {
   rarity: Rarity;
+  onlyOne?: boolean;
 }
 
-export default function RarityList({ rarity }: Props) {
+export default function RarityList({ rarity, onlyOne }: Props) {
   return (
     <ul className={classNames("rarity-list", `rarity-list--${rarityString(rarity)}`)}>
-      {Array.from({ length: rarity }, (_, i) => (
+      {Array.from({ length: onlyOne ? 1 : rarity }, (_, i) => (
         <li key={i} className="rarity-list__item">
           <Star />
         </li>