+++ src/pages/Data/Characters/{characterName}/Character.tsx
@@ -8,5 +8,6 @@ export default function Character() {
     showAscensionSection
     showCharacterPlaystyle
     showSignatureWeapon
+    showRecommendedWeapons
   />} />;
 }